Remove oscilloscope.py script and update TODO list
This commit removes the standalone oscilloscope.py script, which has been replaced by the more comprehensive GUI implementation. The TODO list has been updated to reflect completed tasks and add new improvement suggestions for the oscilloscope data viewer project.
