+++ TODO.md
@@ -1,12 +1,15 @@
 # TODO
 
 - [ ] Refactor the code for better modularity and maintainability.
-- [ ] Fix the CSV selector/viewer not being able to be seen unless fullscreen.
-- [ ] Implement actual cursor measurements.
-- [ ] Improve cursors usage
+- [x] Fix the CSV selector/viewer not being able to be seen unless fullscreen.
+- [x] Implement actual cursor measurements.
+- [ ] Fix cursor measurements not updating correctly after placing second cursor or during dragging.
+- [x] Improve cursors usage
 - [ ] Add dark/light mode for the plots and UI.
+- [x] Fix channel toggling functionality.
 
 # Future:
+
 - [ ] Implement zoom and pan functionality for detailed analysis.
 - [ ] Add support for saving plots as images (e.g., PNG, JPG).
 - [ ] Implement basic signal processing functions (e.g., FFT).
